+++ modules/tinymce/src/plugins/lists/main/ts/core/Util.ts
@@ -1,4 +1,4 @@
-import { Arr, Fun } from '@ephox/katamari';
+import { Arr } from '@ephox/katamari';
 
 import Editor from 'tinymce/core/api/Editor';
 import { NodeChangeEvent } from 'tinymce/core/api/EventTypes';
@@ -27,8 +27,8 @@ const isWithinNonEditableList = (editor: Editor, element: Element | null): boole
   return isWithinNonEditable(editor, parentList);
 };
 
-const hasNonEditableBlocksSelected = (editor: Editor): boolean =>
-  Arr.exists(editor.selection.getSelectedBlocks(), Fun.not(editor.dom.isEditable));
+const isInNonEditableRoot = (editor: Editor): boolean =>
+  editor.getBody().getAttribute('contenteditable') === 'false';
 
 const setNodeChangeHandler = (editor: Editor, nodeChangeHandler: (e: NodeChangeEvent) => void): () => void => {
   const initialNode = editor.selection.getNode();
@@ -46,6 +46,6 @@ export {
   inList,
   selectionIsWithinNonEditableList,
   isWithinNonEditableList,
-  hasNonEditableBlocksSelected,
+  isInNonEditableRoot,
   setNodeChangeHandler
 };