+++ modules/tinymce/CHANGELOG.md
@@ -82,6 +82,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed the constrained bounds calculation for dismissal of toolbar when using `toolbar_location: 'bottom'`. #TINY-9718
 - The floating toolbar did not occupy the entire available space when the page had a flexbox layout. #TINY-9847
 - Fixed backspace and delete keypresses within details elements. #TINY-9884
+- `InsertOrderedList` did not work if there was a `contenteditable=false` element with selection. #TINY-9823
 
 ## 6.4.2 - 2023-04-26